+++ src/core/client/views/hud/controllers/phoneController.ts
@@ -1,15 +1,17 @@
 import * as alt from 'alt-client';
 import { BaseHUD } from '../hud';
-import { PhoneEventList } from '../../../../shared/enums/phoneEvents';
+import { PhoneEventList, PhoneEvents } from '../../../../shared/enums/phoneEvents';
 import { isAnyMenuOpen } from '../../../utility/menus';
 import { VehicleAppController } from './apps/vehicleApp';
 import { BankAppController } from './apps/bankApp';
+import { DealershipAppController } from './apps/dealershipApp';
 
 export class PhoneController {
     static initializeApps() {
         // Initialize Apps
         VehicleAppController.init();
         BankAppController.init();
+        DealershipAppController.init();
     }
 
     static togglePhone() {
@@ -18,6 +20,15 @@ export class PhoneController {
         }
 
         BaseHUD.view.emit('phone:Toggle');
+    }
+
+    static updateTime(hour: number, minute: number) {
+        if (!BaseHUD.view) {
+            return;
+        }
+
+        BaseHUD.view.emit('phone:SetData', 'hour', hour);
+        BaseHUD.view.emit('phone:SetData', 'minute', minute);
     }
 
     /**
@@ -30,7 +41,12 @@ export class PhoneController {
      * @memberof PhoneController
      */
     static routeFromPhone(name: string, ...args: any[]) {
-        const event = PhoneEventList.find((event) => event.name === name);
+        alt.log(name);
+        alt.log(args);
+
+        alt.log(JSON.stringify(PhoneEventList));
+
+        const event = PhoneEventList.find((event) => event.name.toLowerCase().includes(name.toLowerCase()));
 
         if (!event) {
             alt.logError(`No event for ${name} for Phone Application.`);