+++ src/core/shared/enums/phoneEvents.ts
@@ -6,12 +6,15 @@ interface PhoneEvent {
 export const PhoneEvents = {
     EVENT: { name: 'phone:Event' },
     ATM_TRANSFER: { name: 'phone:ATM:Transfer', isServer: true },
-    ATM_PROCESS: { name: 'phone:ATM:Process' }
+    ATM_PROCESS: { name: 'phone:ATM:Process' },
+    DEALERSHIP_BUY: { name: 'phone:Dealership:Buy', isServer: true }
 };
 
 export const PhoneEventList: Array<PhoneEvent> = [
     //
     PhoneEvents.EVENT,
     PhoneEvents.ATM_TRANSFER,
-    PhoneEvents.ATM_PROCESS
+    PhoneEvents.ATM_PROCESS,
+    PhoneEvents.DEALERSHIP_BUY
+    // Add New Phone App Events to Serverside Here
 ];